Share prefix trie matcher with segmentation
Move the recursive UTF-8 prefix trie out of Conversion.cpp into a reusable PrefixMatch helper and wire both conversion and max-match segmentation through it.

The helper still preserves DictGroup ordering by building one table per child dictionary, so the first dictionary with a matching prefix wins exactly as before. Each table performs longest-prefix matching inside that dictionary, storing the matched key for segmentation and the default value for conversion.

This avoids calling the Marisa common-prefix search from the segmentation hot loop. On the zuozhuan long-text benchmark this raised s2t throughput from the previous recursive-conversion-only result of about 12.54 MB/s to 29.52 MB/s, and s2twp from about 10.04 MB/s to 18.86 MB/s.

Verified with: bazel test //src:conversion_test //src:conversion_chain_test //src:max_match_segmentation_test //src:dict_group_test //test/golden:golden_convert_test --test_output=errors; bazel test //src/... //test/... --test_output=errors; cmake --build build/perf --target performance -j8.
