+++ src/BUILD.bazel
@@ -127,8 +127,7 @@ cc_library(
     hdrs = ["Conversion.hpp"],
     deps = [
         ":common",
-        ":dict",
-        ":lexicon",
+        ":prefix_match",
         ":segmentation",
     ],
 )
@@ -341,7 +340,20 @@ cc_library(
     deps = [
         ":common",
         ":dict_group",
+        ":prefix_match",
         ":segmentation",
+    ],
+)
+
+cc_library(
+    name = "prefix_match",
+    srcs = ["PrefixMatch.cpp"],
+    hdrs = ["PrefixMatch.hpp"],
+    deps = [
+        ":common",
+        ":dict",
+        ":lexicon",
+        ":utf8_util",
     ],
 )
